+++ test/distributed/transactions_test.rb
@@ -74,10 +74,10 @@ class TestDistributedTransactions < Minitest::Test
     r.watch("{qux}foo", "{qux}bar", "{qux}baz") do
       assert_equal '1', r.get("{qux}baz")
 
-      result = r.multi do
-        r.incrby("{qux}foo", 3)
-        r.incrby("{qux}bar", 6)
-        r.incrby("{qux}baz", 9)
+      result = r.multi do |transaction|
+        transaction.incrby("{qux}foo", 3)
+        transaction.incrby("{qux}bar", 6)
+        transaction.incrby("{qux}baz", 9)
       end
 
       assert_equal [3, 6, 10], result