+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 # Unreleased 5.0.0
 
+- Removed the deprecated `pipelined` and `multi` signature. Commands now MUST be called on the block argument, not the original redis instance.
 - Removed `Redis.current`. You shouldn't assume there is a single global Redis connection, use a connection pool instead,
   and libaries using Redis should accept a Redis instance (or connection pool) as a config. E.g. `MyLibrary.redis = Redis.new(...)`.
 - Removed the `synchrony` driver.